Move logic functions to own file
