+++ modules/public/views.py
@@ -3,7 +3,7 @@ import datetime
 import flask
 import humanize
 import modules.public.api as api
-import pycountry
+import modules.public.logic as logic
 import re
 from dateutil import parser, relativedelta
 from modules.exceptions import (
@@ -15,158 +15,7 @@ from modules.exceptions import (
     ApiConnectionError
 )
 from math import floor
-from urllib.parse import parse_qs, urlparse, quote_plus
-
-
-def calculate_colors(countries, max_users):
-    for country_code in countries:
-        countries[country_code]['color_rgb'] = [
-            calculate_color(
-                countries[country_code]['percentage_of_users'],
-                max_users,
-                8,
-                229
-            ),
-            calculate_color(
-                countries[country_code]['percentage_of_users'],
-                max_users,
-                64,
-                245
-            ),
-            calculate_color(
-                countries[country_code]['percentage_of_users'],
-                max_users,
-                129,
-                223
-            )
-        ]
-
-
-def calculate_color(thisCountry, maxCountry, maxColor, minColor):
-    countryFactor = float(thisCountry)/maxCountry
-    colorRange = maxColor - minColor
-    return int(colorRange*countryFactor+minColor)
-
-
-def normalize_metrics(geodata):
-    users_by_country = {}
-    max_users = 0.0
-    for country_counts in geodata:
-        country_code = country_counts['name']
-        users_by_country[country_code] = {}
-        counts = []
-        for daily_count in country_counts['values']:
-            if daily_count is not None:
-                counts.append(daily_count)
-
-        if len(counts) > 0:
-            users_by_country[country_code]['number_of_users'] = (
-                sum(counts)
-            )
-            users_by_country[country_code]['percentage_of_users'] = (
-                sum(counts) / len(counts)
-            )
-        else:
-            users_by_country[country_code]['number_of_users'] = 0
-            users_by_country[country_code]['percentage_of_users'] = 0
-
-        if max_users < users_by_country[country_code]['percentage_of_users']:
-            max_users = users_by_country[country_code]['percentage_of_users']
-
-    calculate_colors(users_by_country, max_users)
-
-    return users_by_country
-
-
-def build_country_info(users_by_country, display_number_users=False):
-    country_data = {}
-    for country in pycountry.countries:
-        country_info = users_by_country.get(country.alpha_2)
-        number_of_users = 0
-        percentage_of_users = 0
-        color_rgb = [247, 247, 247]
-        if country_info is not None:
-            if display_number_users:
-                number_of_users = country_info['number_of_users'] or 0
-            percentage_of_users = country_info['percentage_of_users'] or 0
-            color_rgb = country_info['color_rgb'] or [247, 247, 247]
-
-        # Use common_name if available to be less political offending (#310)
-        try:
-            country_name = country.common_name
-        except AttributeError:
-            country_name = country.name
-
-        country_data[country.numeric] = {
-            'name': country_name,
-            'code': country.alpha_2,
-            'percentage_of_users': percentage_of_users,
-            'color_rgb': color_rgb
-        }
-
-        if display_number_users:
-            country_data[country.numeric]['number_of_users'] = number_of_users
-
-    return country_data
-
-
-def normalize_searched_snaps(search_results):
-    return (
-        search_results['_embedded']['clickindex:package']
-        if '_embedded' in search_results
-        else []
-    )
-
-
-def get_pages_details(links):
-    links_result = {}
-
-    if('first' in links):
-        links_result['first'] = convert_limit_offset_to_size_page(
-            links['first']['href']
-        )
-
-    if('last' in links):
-        links_result['last'] = convert_limit_offset_to_size_page(
-            links['last']['href']
-        )
-
-    if('next' in links):
-        links_result['next'] = convert_limit_offset_to_size_page(
-            links['next']['href']
-        )
-
-    if('prev' in links):
-        links_result['prev'] = convert_limit_offset_to_size_page(
-            links['prev']['href']
-        )
-
-    if('self' in links):
-        links_result['self'] = convert_limit_offset_to_size_page(
-            links['self']['href']
-        )
-
-    return links_result
-
-
-def convert_limit_offset_to_size_page(link):
-    url_parsed = urlparse(link)
-    host_url = (
-        "{base_url}"
-        "?q={q}&limit={limit}&offset={offset}"
-    )
-
-    url_queries = parse_qs(url_parsed.query)
-    q = url_queries['q'][0]
-    size = int(url_queries['size'][0])
-    page = int(url_queries['page'][0])
-
-    return host_url.format(
-        base_url=flask.request.base_url,
-        q=q,
-        limit=size,
-        offset=size*(page-1)
-    )
+from urllib.parse import urlparse, quote_plus
 
 
 def _handle_errors(api_error: ApiError):
@@ -195,7 +44,9 @@ def homepage():
     error_info = {}
     status_code = 200
     try:
-        featured_snaps = normalize_searched_snaps(api.get_featured_snaps())
+        featured_snaps = logic.normalize_searched_snaps(
+            api.get_featured_snaps()
+        )
     except ApiError as api_error:
         status_code, error_info = _handle_errors(api_error)
 
@@ -211,7 +62,9 @@ def store():
     error_info = {}
     status_code = 200
     try:
-        featured_snaps = normalize_searched_snaps(api.get_featured_snaps())
+        featured_snaps = logic.normalize_searched_snaps(
+            api.get_featured_snaps()
+        )
     except ApiError as api_error:
         status_code, error_info = _handle_errors(api_error)
 
@@ -228,7 +81,9 @@ def snaps():
     error_info = {}
     status_code = 200
     try:
-        promoted_snaps = normalize_searched_snaps(api.get_promoted_snaps())
+        promoted_snaps = logic.normalize_searched_snaps(
+            api.get_promoted_snaps()
+        )
     except ApiError as api_error:
         status_code, error_info = _handle_errors(api_error)
 
@@ -260,8 +115,8 @@ def search_snap():
             page
         )
 
-        normalize_results = normalize_searched_snaps(searched_results)
-        links = get_pages_details(
+        normalize_results = logic.normalize_searched_snaps(searched_results)
+        links = logic.get_pages_details(
             (
                 searched_results['_links']
                 if '_links' in searched_results
@@ -355,10 +210,10 @@ def snap_details(snap_name):
             metrics_query_json
         )
 
-        users_by_country = normalize_metrics(
+        users_by_country = logic.normalize_metrics(
             metrics_response[0]['series']
         )
-        country_data = build_country_info(users_by_country)
+        country_data = logic.build_country_info(users_by_country)
     except ApiError as api_error:
         status_code, error_info = _handle_errors(api_error)
